Add files via upload
This project is a comprehensive DBMS assignment allotted by our teacher in university as a part of internalsthat involves creating a database system for the Indian National Health Mission. It includes an Entity-Relationship (ER) diagram, database creation, and implementation of various SQL concepts such as queries, joins, constraints, and functions. The project demonstrates how SQL can be utilized to manage health-related data efficiently. All code and outputs are provided to illustrate the SQL functionalities within this context.